requirements.txt
db-dtypes==1.0.4
gcsfs==2022.11.0
google-cloud-bigquery==3.3.6
google-cloud-storage==2.6.0
pandas==1.3.5
scikit-learn==1.0.2
tensorflow==2.9.0